Сделал 2 задание 3 практики(добавил зависимости для автоматического создания Open-Api спецификации)
