+++ bot/src/main/java/ru/tinkoff/edu/java/bot/BotApplication.java
@@ -8,9 +8,9 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 @SpringBootApplication
 @EnableConfigurationProperties(ApplicationConfig.class)
 public class BotApplication {
-public static void main(String[] args) {
-        var ctx = SpringApplication.run(BotApplication.class, args);
-        ApplicationConfig config = ctx.getBean(ApplicationConfig.class);
-        System.out.println(config);
+        public static void main(String[] args) {
+                var ctx = SpringApplication.run(BotApplication.class, args);
+                ApplicationConfig config = ctx.getBean(ApplicationConfig.class);
+                System.out.println(config);
         }
-        }
+}
