+++ bot/src/main/java/ru/tinkoff/edu/java/bot/controller/UpdateController.java
@@ -1,6 +1,7 @@
 package ru.tinkoff.edu.java.bot.controller;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -11,7 +12,10 @@ import ru.tinkoff.edu.java.bot.dto.LinkUpdateResponse;
 
 @Validated
 @RestController
-@RequestMapping("/updates")
+@RequestMapping(path = "/updates",
+        consumes = MediaType.APPLICATION_JSON_VALUE,
+        produces = MediaType.APPLICATION_JSON_VALUE
+)
 public class UpdateController {
 
     @PostMapping
