+++ bot/src/main/resources/application.yml
@@ -1,2 +1,7 @@
 bot:
-  test: Helloworld!!
+  test: "Helloworld!!"
+
+springdoc:
+  swagger-ui:
+    enabled: true
+    path: /swagger-ui
