+++ pom.xml
@@ -23,6 +23,7 @@
     <maven.compiler.target>1.7</maven.compiler.target>
     <spring-cloud.version>2022.0.0</spring-cloud.version>
     <spring-boot.version>3.0.1</spring-boot.version>
+    <springdoc.version>2.0.2</springdoc.version>
   </properties>
   <dependencyManagement>
     <dependencies>
@@ -41,6 +42,13 @@
       <type>pom</type>
       <scope>import</scope>
     </dependency>
+
+      <dependency>
+        <groupId>org.springdoc</groupId>
+        <artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
+        <version>${springdoc.version}</version>
+      </dependency>
+
     </dependencies>
   </dependencyManagement>
