+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/controller/TgChatController.java
@@ -1,6 +1,7 @@
 package ru.tinkoff.edu.java.scrapper.controller;
 
 import jakarta.validation.constraints.PositiveOrZero;
+import org.springframework.http.MediaType;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,7 +14,10 @@ import java.util.NoSuchElementException;
 
 @Validated
 @RestController
-@RequestMapping("/tg-chat")
+@RequestMapping(path = "/tg-chat",
+        consumes = MediaType.APPLICATION_JSON_VALUE,
+        produces = MediaType.APPLICATION_JSON_VALUE
+)
 public class TgChatController {
     @PostMapping("/{id}")
     public LinkResponse register(@PathVariable @PositiveOrZero long id) {
